Externalize some scripts, add some more stuff, seems to be getting better
